requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
idna==3.2
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
pymongo==3.12.0
requests==2.26.0
typing-extensions==3.10.0.0
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
zipp==3.5.0
